+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
+- Support for multiple profiles. You can now use multiple Invidious accounts.
 - When a video error dialog shows, a timer with 10 seconds starts to play the next video
 
 ### Fixed